requirements.txt
requests-oauthlib==1.3.1
google-cloud-vision==3.4.0
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
